refactor PermissionSelectorを切り出し
